+++ Model/Resolver/CreditMemo/CreditMemoItems.php
@@ -161,7 +161,7 @@ class CreditMemoItems extends SourceCreditMemoItems
             $discounts[] = [
                 'label' => $associatedOrder->getDiscountDescription() ?? __('Discount'),
                 'amount' => [
-                    'value' => abs((int)$creditmemoItem->getDiscountAmount()),
+                    'value' => abs((float)$creditmemoItem->getDiscountAmount()),
                     'currency' => $associatedOrder->getOrderCurrencyCode()
                 ]
             ];
